refactor(ThreadController): remove apiCall
