+++ app/src/main/java/com/infomaniak/mail/ui/main/thread/ThreadViewModel.kt
@@ -24,15 +24,10 @@ import androidx.lifecycle.liveData
 import androidx.lifecycle.viewModelScope
 import com.infomaniak.mail.data.LocalSettings
 import com.infomaniak.mail.data.api.ApiRepository
-import com.infomaniak.mail.data.cache.RealmDatabase
-import com.infomaniak.mail.data.cache.mailboxContent.DraftController
 import com.infomaniak.mail.data.cache.mailboxContent.MessageController
 import com.infomaniak.mail.data.cache.mailboxContent.ThreadController
-import com.infomaniak.mail.data.cache.mailboxInfo.MailboxController
 import com.infomaniak.mail.data.models.Mailbox
 import com.infomaniak.mail.data.models.message.Message
-import com.infomaniak.mail.data.models.thread.Thread
-import com.infomaniak.mail.utils.AccountUtils
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.launch
@@ -47,51 +42,6 @@ class ThreadViewModel(application: Application) : AndroidViewModel(application) 
 
     fun messagesLive(threadUid: String) = liveData(Dispatchers.IO) {
         ThreadController.getThread(threadUid)?.messages?.asFlow()?.asLiveData()?.let { emitSource(it) }
-    }
-
-    fun openThread(threadUid: String) = liveData(Dispatchers.IO) {
-
-        val mailbox = MailboxController.getMailbox(AccountUtils.currentUserId, AccountUtils.currentMailboxId) ?: run {
-            emit(null)
-            return@liveData
-        }
-
-        val thread = ThreadController.getThread(threadUid) ?: run {
-            emit(null)
-            return@liveData
-        }
-
-        val expandedList = thread.messages.mapIndexed { index, message ->
-            !message.isDraft && (!message.seen || index == thread.messages.lastIndex)
-        }
-        emit(expandedList)
-
-        fetchIncompleteMessages(thread)
-
-        if (thread.unseenMessagesCount > 0) {
-            ThreadController.markAsSeen(thread, mailbox.uuid)
-            MessageController.fetchCurrentFolderMessages(mailbox, thread.folderId, localSettings.threadMode)
-        }
-    }
-
-    private fun fetchIncompleteMessages(thread: Thread) {
-        RealmDatabase.mailboxContent().writeBlocking {
-            thread.messages.forEach { localMessage ->
-                if (!localMessage.fullyDownloaded) {
-                    ApiRepository.getMessage(localMessage.resource).data?.also {
-                        it.messageIds = localMessage.messageIds
-
-                        // If we've already got this Message's Draft beforehand, we need to save
-                        // its `draftLocalUuid`, otherwise we'll lose the link between them.
-                        if (it.isDraft) it.draftLocalUuid = DraftController.getDraftByMessageUid(it.uid, realm = this)?.localUuid
-
-                        it.fullyDownloaded = true
-
-                        MessageController.upsertMessage(it, realm = this)
-                    }
-                }
-            }
-        }
     }
 
     fun deleteDraft(message: Message, threadUid: String, mailbox: Mailbox) = viewModelScope.launch(Dispatchers.IO) {
